# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a table top game.",
        "title": "Table Top API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/enemy_moves": {
            "get": {
                "description": "List all enemy moves, optionally filtered by EnemyID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EnemyMoves"
                ],
                "summary": "List all enemy moves",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enemy ID",
                        "name": "EnemyID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EnemyMove"
                            }
                        }
                    }
                }
            }
        },
        "/enemy_moves/{enemyId}": {
            "post": {
                "description": "Create a new enemy move",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EnemyMoves"
                ],
                "summary": "Create a new enemy move",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enemy ID",
                        "name": "enemyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "EnemyMove",
                        "name": "enemyMove",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enemy_moves.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EnemyMove"
                        }
                    }
                }
            }
        },
        "/enemy_moves/{id}": {
            "get": {
                "description": "Get an enemy move by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EnemyMoves"
                ],
                "summary": "Get an enemy move by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EnemyMove ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EnemyMove"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing enemy move",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EnemyMoves"
                ],
                "summary": "Update an existing enemy move",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EnemyMove ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "EnemyMove",
                        "name": "enemyMove",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enemy_moves.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EnemyMove"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an enemy move by ID",
                "tags": [
                    "EnemyMoves"
                ],
                "summary": "Delete an enemy move by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EnemyMove ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/locations": {
            "get": {
                "description": "List all locations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "List all locations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Location"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Create a new location",
                "parameters": [
                    {
                        "description": "Location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/locations.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    }
                }
            }
        },
        "/locations/{id}": {
            "get": {
                "description": "Get a location by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Get a location by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Update a location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/locations.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a location by ID",
                "tags": [
                    "Locations"
                ],
                "summary": "Delete a location by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enemy_moves.CreateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Close attack"
                },
                "range_from": {
                    "type": "integer",
                    "example": 1
                },
                "range_to": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "enemy_moves.UpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "range_from": {
                    "type": "integer"
                },
                "range_to": {
                    "type": "integer"
                }
            }
        },
        "locations.CreateRequest": {
            "type": "object",
            "properties": {
                "danger_level": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "monsters_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "locations.UpdateRequest": {
            "type": "object",
            "properties": {
                "danger_level": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "monsters_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.EnemyMove": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Slash attack"
                },
                "enemy_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "range_from": {
                    "type": "integer",
                    "example": 1
                },
                "range_to": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "danger_level": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "monsters_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}